Title: 사칙연산 함수, Time: 123ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240520/사칙연산 함수/quadratic-operations-function.py
@@ -4,7 +4,7 @@ c = int(c)
 
 def o_judg(o):
     if o == "/":
-        return a / c
+        return int(a / c)
     if o == '*':
         return a * c
     if o == '+':
